THIS SHOULD WORK!
Seems to have something to do with pattern matching.  Given a tuple (A,
B), if we don't care about B it matches properly regardless of the value
of A.  This does not seem to be the case the other way around.

I'm committing this even if causes a totally different test to what we
added to fail.  Will figure this out later.